Backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pillow==11.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
